AWS DAGMAN WORKFLOW/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==21.3.1
setuptools==60.2.0
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.11.0
pyodbc
scikit-learn
pandas==1.4.4
numpy<1.24
email-to
sqlalchemy
htcondor==9.1.3
classad
cython
pegasuspy
openpyxl
defusedxml